Use new Docker mount options
The latest Docker allows us to specify the :cached and :delegated
options.  :cached means that the host has an authorative view over
some files, :delegated means the container has an authorative view.
The files will eventually be consistent, but you can expect only the
host or the container to edit these files.  If there would be a
conflict, the previously described 'owner' wins.