Optionally mount only used node_modules
Although not the target of this code, this solution helps limit the
amount of node_modules.  This may be important on MacOS, as adding
mounted volumes seems to add latency to filesystem access.

The codepath for this solution was constructed in the hopes of
supporting nested linked node_modules dependencies.  It seems that the
node_modules are seemingly stored in the container correctly but that
the backing pipeline doesn't support this case, hence the current
solution will not cover it.  It may, however, be that we use this more
intelligent solution to develop a workaround for said case.