Allow direct linking of local node modules
Local linking allows us to circumvent some build issues when using
deeply nested addons under development.  In the regular way of
working, these addons all need to be installed on the top-level
application.  This is not always the desired scenario.  Although we
still support that way of working, this commit allows us to be
slightly smarter.

Local linking might also make the system slightly faster.  Though we
haven't run any benchmarks and don't have the faintest idea whether
that is an issue on Docker for Mac or not, it is an interesting
thought.