Extract support-scripts into separate file
The support scripts are extracted into a separate file.  These scripts
are combined when a release is made.  Although the code is still
physically duplicated, updating the source becomes easier.

We could have sourced the file which contains the scripts instead of
injecting them.  Searching online for finding the real name of a
script seemed to cause minor issues.  Hence we've chosen the path
where we inject the sources (for now).