Fix npm_link for MacOS
sed -r doesn't exist on MacOS, sed -E seems to behave similarly for
this query.  sed -E doesn't exist on Linux's sed.  Hence we're left
with the combination of the two.

Indentation has been made consistent with the rest of the file.
Something that was seemingly left in for a while.