Add support for foldernames which contain spaces
Foldernames currently can't contain spaces.  This alters a bunch of
the workings of the script in a way that makes this possible.

Working this out in a different way seemed hard.  The use of eval
for a return value made it comparatively hard to correctly escape the
strings.  As such, we've reworked the passing of variables where it
made sense and work with this different approach now.  With the API
surface of edi having stabilized, I think this makes sense.